Fix #2 partially.
- There is some issue with facebook's login dialog popup window that
causes a blank popup after login with facebook. I tried to fix this with
out any success.
- With this fix, the login should work by following a sequence of steps
noted in the point below. It's ugly as hell, but it works.
- If you close the popup, go back to home screen by clicking on saavn
logo, click login again, click on login with facebook, close the empty
popup, you should be redirected back to home screen with your account
logged in.
- This is some ugly shit. Why can't saavn release an app of their own
already? Why the hell are we having to muck with electron here?